use float for range&scale, add axe annotation

+++ jzy3d-api/src/api/org/jzy3d/maths/Range.java
@@ -2,7 +2,7 @@ package org.jzy3d.maths;
 
 /** For some naming relevance, a simple extention of Scale.*/
 public class Range extends Scale {
-	public Range(double min, double max){
+	public Range(float min, float max){
 		super(min, max);
 	}
