+++ jzy3d-api/src/api/org/jzy3d/maths/algorithms/OutlierRemover.java
@@ -4,6 +4,7 @@ import org.jzy3d.maths.Array;
 import org.jzy3d.maths.Scale;
 import org.jzy3d.maths.Statistics;
 
+// warn : cast (float)
 public class OutlierRemover {
 	public static int[] getOutlierIndices(double[] values, int nVariance){
 		Scale bounds = getInlierBounds(values, nVariance);
@@ -12,7 +13,7 @@ public class OutlierRemover {
 		int k = 0;
 		
 		for(int i=0; i<values.length; i++)
-			if( !bounds.contains(values[i]) )
+			if( !bounds.contains((float)values[i]) )
 				selection[k++] = i;
 		
 		return Array.clone(selection, k);
@@ -25,13 +26,11 @@ public class OutlierRemover {
 		int k = 0;
 		
 		for(int i=0; i<values.length; i++)
-			if( bounds.contains(values[i]) )
+			if( bounds.contains((float)values[i]) )
 				selection[k++] = i;
 		
 		return Array.clone(selection, k);
 	}
-	
-	/************************************************************************/
 	
 	public static double[] getOutlierValues(double[] values, int nVariance){
 		Scale bounds = getInlierBounds(values, nVariance);
@@ -40,7 +39,7 @@ public class OutlierRemover {
 		int k = 0;
 		
 		for(int i=0; i<values.length; i++)
-			if( !bounds.contains(values[i]) )
+			if( !bounds.contains((float)values[i]) )
 				selection[k++] = values[i];
 		
 		return Array.clone(selection, k);
@@ -53,35 +52,29 @@ public class OutlierRemover {
 		int k = 0;
 		
 		for(int i=0; i<values.length; i++)
-			if( bounds.contains(values[i]) )
+			if( bounds.contains((float)values[i]) )
 				selection[k++] = values[i];
 		
 		return Array.clone(selection, k);
 	}
 	
-	/************************************************************************/
-	
 	public static Scale getInlierBounds(double[] values, int nVariance){
 		if(values.length==0)
-			return new Scale(Double.NaN, Double.NaN);
+			return new Scale(Float.NaN, Float.NaN);
 		
 		// Compute stdist: median of distances to median
 		double[] dists = new double[values.length];
-		double med  = Statistics.median(values, true);
-		double mad  = 0;
+		float med  = (float)Statistics.median(values, true);
+		float mad  = 0;
 		
 		for(int i=0; i<values.length; i++)
 			dists[i] = Math.abs( values[i] - med );
-		mad = Statistics.median(dists, true);
+		mad = (float)Statistics.median(dists, true);
 		
 		// Compute the acceptance region bounds
-		double upperBound = med + mad * nVariance;
-		double lowerBound = med - mad * nVariance;
+		float upperBound = med + mad * nVariance;
+		float lowerBound = med - mad * nVariance;
 		
 		return new Scale(lowerBound, upperBound);
 	}
-	
-	/***********************************************************************/
-	
-	
 }
